requirements-dev.txt
alembic==1.12.0
amqp==5.1.1
appdirs==1.4.4
astroid==2.15.6
attrs==23.1.0
autopage==0.5.1
bcrypt==4.0.1
black==23.9.1
build==1.0.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cliff==4.3.0
cmd2==2.4.3
cryptography==41.0.3
debtcollector==2.5.0
decorator==5.1.1
dill==0.3.7
dnspython==2.4.2
docutils==0.20.1
dogpile.cache==1.2.2
eventlet==0.33.3
fasteners==0.18
fixtures==4.1.0
futurist==2.4.1
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
iso8601==2.0.0
isort==5.12.0
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keyring==24.2.0
keystoneauth1==5.3.0
keystonemiddleware==10.4.1
kombu==5.3.2
lazy-object-proxy==1.9.0
logutils==0.3.5
lxml==4.9.3
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.1.0
msgpack==1.0.5
mypy-extensions==1.0.0
ncclient==0.6.13
netaddr==0.8.0
netifaces==0.11.0
neutron==22.0.2
neutron-lib==3.8.0
nh3==0.2.14
openstacksdk==1.5.0
os-client-config==2.1.0
os-ken==2.7.0
os-resource-classes==1.1.0
os-service-types==1.7.0
os-traits==3.0.0
os-vif==3.2.0
osc-lib==2.8.1
oslo.cache==3.5.0
oslo.concurrency==5.2.0
oslo.config==9.2.0
oslo.context==5.2.0
oslo.db==14.0.0
oslo.i18n==6.1.0
oslo.log==5.3.0
oslo.messaging==14.4.0
oslo.metrics==0.7.0
oslo.middleware==5.2.0
oslo.policy==4.2.1
oslo.privsep==3.2.0
oslo.reports==3.1.0
oslo.rootwrap==7.1.0
oslo.serialization==5.2.0
oslo.service==3.2.0
oslo.upgradecheck==2.2.0
oslo.utils==6.2.1
oslo.versionedobjects==3.2.0
osprofiler==4.1.0
ovs==3.1.2
ovsdbapp==2.4.1
packaging==23.1
paramiko==3.3.1
Paste==3.5.3
PasteDeploy==3.0.1
pathspec==0.11.2
pbr==5.11.1
pecan==1.5.1
pkginfo==1.9.6
platformdirs==3.10.0
prettytable==3.9.0
prometheus-client==0.17.1
psutil==5.9.5
pycadf==3.1.1
pycparser==2.21
Pygments==2.16.1
pyinotify==0.9.6
pylint==2.17.5
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyperclip==1.8.2
pyproject_hooks==1.0.0
pyroute2==0.7.9
python-dateutil==2.8.2
python-designateclient==5.3.0
python-keystoneclient==5.1.0
python-neutronclient==11.0.0
python-novaclient==18.4.0
pytz==2023.3.post1
PyYAML==6.0.1
readme-renderer==42.0
referencing==0.30.2
repoze.lru==0.7
requests==2.31.0
requests-toolbelt==1.0.0
requestsexceptions==1.4.0
rfc3986==2.0.0
rich==13.5.2
Routes==2.5.1
rpds-py==0.10.3
SecretStorage==3.3.3
setproctitle==1.3.2
simplejson==3.19.1
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.20
statsd==4.0.1
stevedore==5.1.0
tenacity==8.2.3
testresources==2.0.1
testscenarios==0.5.0
testtools==2.6.0
tomlkit==0.12.1
tooz==4.2.0
twine==4.0.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.6
WebOb==1.8.7
wrapt==1.15.0
yappi==1.4.0
zipp==3.16.2
